requirements.txt
baidu-aip==2.2.4.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-debug-toolbar==1.9.1
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.1.0
pytz==2018.4
requests==2.18.4
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
